remission check folio fix, mvc configuration added

+++ src/main/java/com/rms/services/service/RemissionAdminService.java
@@ -38,7 +38,7 @@ public class RemissionAdminService {
 
 	@Autowired
 	ProductRepository productRepository;
-	
+
 	@Autowired
 	AreaRepository areaRepository;
 
@@ -46,21 +46,18 @@ public class RemissionAdminService {
 
 		RemissionVO remissionVO = new RemissionVO();
 		Optional<Remission> result = remissionRepository.findById(remissionId);
-		
-		
 
 		if (result.isPresent()) {
-				
+
 			Remission remission = result.get();
-			
+
 			Optional<Area> area = areaRepository.findById(remission.getDestinationId());
 
 			BeanUtils.copyProperties(remission, remissionVO);
 
-			if(area.isPresent()) {
-				remissionVO.setClient(area.get().getClientName());	
+			if (area.isPresent()) {
+				remissionVO.setClient(area.get().getClientName());
 			}
-			
 
 			remissionVO.setProducts(getRemissionProducts(remission));
 
@@ -85,10 +82,13 @@ public class RemissionAdminService {
 	}
 
 	public RemissionVO createRemission(@Valid RemissionVO remisionRequest) {
-		
-		long noRemision = remissionRepository.findByFolio(remisionRequest.getFolio()).stream()
-		.filter(rm -> !rm.getId().equalsIgnoreCase(remisionRequest.getId())).count();
-		
+
+		long noRemision = 0;
+		if (!remisionRequest.getFolio().isBlank()) {
+			noRemision = remissionRepository.findByFolio(remisionRequest.getFolio()).stream()
+					.filter(rm -> !rm.getId().equalsIgnoreCase(remisionRequest.getId())).count();
+		}
+
 		Remission remission;
 
 		final List<Product> productList = new ArrayList<>();
@@ -107,31 +107,25 @@ public class RemissionAdminService {
 		// needs to be created.
 		if (remisionRequest.getId() == null || remisionRequest.getId().isBlank()
 				|| remisionRequest.getId().equalsIgnoreCase("0")) {
-			
-			if(noRemision != 0) {
+
+			if (noRemision != 0) {
 				throw new AccessPointException("El Folio proporcionado ya existe en otra remisión");
 			}
-				
+
 			remission = Remission.builder().folio(remisionRequest.getFolio())
 					.destinationId(remisionRequest.getDestinationId())
-					.areaDescription(remisionRequest.getAreaDescription())
-					.userId("rogelio")
-					.greatTotal(remisionRequest.getGreatTotal())
-					.total(remisionRequest.getTotal())
-					.formatId(remisionRequest.getFormatId())
-					.products(productList)
-					.expeditionDate(remisionRequest.getExpeditionDate())
-					.status("Activa")
-					.creationDate(new Date())
-					.observations(remisionRequest.getObservations())
-					.userId("1").build();
+					.areaDescription(remisionRequest.getAreaDescription()).userId("rogelio")
+					.greatTotal(remisionRequest.getGreatTotal()).total(remisionRequest.getTotal())
+					.formatId(remisionRequest.getFormatId()).products(productList)
+					.expeditionDate(remisionRequest.getExpeditionDate()).status("Activa").creationDate(new Date())
+					.observations(remisionRequest.getObservations()).userId("1").build();
 
 			remission.setProducts(productList);
 			// Save the new Complex entity to the database.
 			remission = remissionRepository.save(remission);
 		} else {
-			
-			if(noRemision != 0) {
+
+			if (noRemision != 0) {
 				throw new AccessPointException("El Folio proporcionado ya existe en otra remisión");
 			}
 			// If the ID is not null or blank, attempt to find the existing Complex.
@@ -159,51 +153,51 @@ public class RemissionAdminService {
 		return remisionRequest;
 	}
 
-
 	public List<RemissionVO> getFilteredRemission(FiltersVO filters) {
-	    List<RemissionVO> remissionVOList = new ArrayList<>();
-	    List<Remission> remissions;
+		List<RemissionVO> remissionVOList = new ArrayList<>();
+		List<Remission> remissions;
 
-	    // Optimize conditional structure for readability and efficiency
-	    if (!Optional.ofNullable(filters.getFolio()).orElse("").isEmpty() && !Optional.ofNullable(filters.getDestinationId()).orElse("").isEmpty()) {
-	        remissions = remissionRepository.findByDestinationIdAndFolioId(filters.getFolio(), filters.getDestinationId());
-	    } else if (!Optional.ofNullable(filters.getFolio()).orElse("").isEmpty()) {
-	        remissions = remissionRepository.findByFolio(filters.getFolio());
-	    } else if (!Optional.ofNullable(filters.getDestinationId()).orElse("").isEmpty()) {
-	        remissions = remissionRepository.findByDestinationIdOrderByExpeditionDate(filters.getDestinationId());
-	    } else {
-	        remissions = new ArrayList<>();
-	    }
+		// Optimize conditional structure for readability and efficiency
+		if (!Optional.ofNullable(filters.getFolio()).orElse("").isEmpty()
+				&& !Optional.ofNullable(filters.getDestinationId()).orElse("").isEmpty()) {
+			remissions = remissionRepository.findByDestinationIdAndFolioId(filters.getFolio(),
+					filters.getDestinationId());
+		} else if (!Optional.ofNullable(filters.getFolio()).orElse("").isEmpty()) {
+			remissions = remissionRepository.findByFolio(filters.getFolio());
+		} else if (!Optional.ofNullable(filters.getDestinationId()).orElse("").isEmpty()) {
+			remissions = remissionRepository.findByDestinationIdOrderByExpeditionDate(filters.getDestinationId());
+		} else {
+			remissions = new ArrayList<>();
+		}
 
-	    // Filter by date range only if no other filters have been applied
-	    if (remissions.isEmpty() && filters.getDateInit() != null && filters.getDateEnd() != null) {
-	        remissions = remissionRepository.findByInitAndEndExpeditionDate(filters.getDateInit(), filters.getDateEnd());
-	    }
+		// Filter by date range only if no other filters have been applied
+		if (remissions.isEmpty() && filters.getDateInit() != null && filters.getDateEnd() != null) {
+			remissions = remissionRepository.findByInitAndEndExpeditionDate(filters.getDateInit(),
+					filters.getDateEnd());
+		}
 
-	    remissions.forEach(remission -> {
-	        RemissionVO remissionVO = new RemissionVO();
-	        BeanUtils.copyProperties(remission, remissionVO);
+		remissions.forEach(remission -> {
+			RemissionVO remissionVO = new RemissionVO();
+			BeanUtils.copyProperties(remission, remissionVO);
 
-	        try {
-	            // Streamline the grouping and concatenation of product descriptions
-	            String partidaDescriptions = remission.getProducts().stream()
-	                    .map(Product::getPartidaDescription)
-	                    .distinct()
-	                    .collect(Collectors.joining(", "));
-	            remissionVO.setRemissionPartidas(partidaDescriptions);
-	        } catch (Exception e) {
-	            // Log the exception for better error tracking
-	            e.printStackTrace();
-	        }
+			try {
+				// Streamline the grouping and concatenation of product descriptions
+				String partidaDescriptions = remission.getProducts().stream().map(Product::getPartidaDescription)
+						.distinct().collect(Collectors.joining(", "));
+				remissionVO.setRemissionPartidas(partidaDescriptions);
+			} catch (Exception e) {
+				// Log the exception for better error tracking
+				e.printStackTrace();
+			}
 
-	        remissionVO.setDestinationId(remission.getDestinationId());
-	        remissionVO.add(linkTo(methodOn(RemissionAdminController.class).getOneRemission(remission.getId()))
-	                .withSelfRel().withName("info"));
+			remissionVO.setDestinationId(remission.getDestinationId());
+			remissionVO.add(linkTo(methodOn(RemissionAdminController.class).getOneRemission(remission.getId()))
+					.withSelfRel().withName("info"));
 
-	        remissionVOList.add(remissionVO);
-	    });
+			remissionVOList.add(remissionVO);
+		});
 
-	    return remissionVOList;
+		return remissionVOList;
 	}
 
 	List<ProductVO> getRemissionProducts(Remission remission) {
@@ -251,8 +245,8 @@ public class RemissionAdminService {
 		Remission newRemission = Remission.builder().folio(remisionRequest.getFolio())
 				.destinationId(remission.getDestinationId()).userId("rogelio").total(remission.getTotal())
 				.formatId(remission.getFormatId()).products(productList).expeditionDate(remission.getExpeditionDate())
-				.areaDescription(remission.getAreaDescription())
-				.creationDate(new Date()).status("Activa").originFolio(remission.getFolio()).userId("1").build();
+				.areaDescription(remission.getAreaDescription()).creationDate(new Date()).status("Activa")
+				.originFolio(remission.getFolio()).userId("1").build();
 
 		newRemission.setProducts(productList);
 		// Save the new Complex entity to the database.
@@ -262,114 +256,112 @@ public class RemissionAdminService {
 		BeanUtils.copyProperties(newRemission, remisionRequest);
 		return remisionRequest;
 	}
-	
+
 	public List<ProductDetailedVO> getFilteredRemissionDetailed(FiltersVO filters) {
-	    List<ProductDetailedVO> productList = new ArrayList<>();
-	    List<Remission> remissions;
+		List<ProductDetailedVO> productList = new ArrayList<>();
+		List<Remission> remissions;
 
-	    // Optimize conditional structure for readability and efficiency
-	    if (!Optional.ofNullable(filters.getFolio()).orElse("").isEmpty() && !Optional.ofNullable(filters.getDestinationId()).orElse("").isEmpty()) {
-	        remissions = remissionRepository.findByDestinationIdAndFolioId(filters.getFolio(), filters.getDestinationId());
-	    } else if (!Optional.ofNullable(filters.getFolio()).orElse("").isEmpty()) {
-	        remissions = remissionRepository.findByFolio(filters.getFolio());
-	    } else if (!Optional.ofNullable(filters.getDestinationId()).orElse("").isEmpty()) {
-	        remissions = remissionRepository.findByDestinationIdOrderByExpeditionDate(filters.getDestinationId());
-	    } else {
-	        remissions = new ArrayList<>();
-	    }
+		// Optimize conditional structure for readability and efficiency
+		if (!Optional.ofNullable(filters.getFolio()).orElse("").isEmpty()
+				&& !Optional.ofNullable(filters.getDestinationId()).orElse("").isEmpty()) {
+			remissions = remissionRepository.findByDestinationIdAndFolioId(filters.getFolio(),
+					filters.getDestinationId());
+		} else if (!Optional.ofNullable(filters.getFolio()).orElse("").isEmpty()) {
+			remissions = remissionRepository.findByFolio(filters.getFolio());
+		} else if (!Optional.ofNullable(filters.getDestinationId()).orElse("").isEmpty()) {
+			remissions = remissionRepository.findByDestinationIdOrderByExpeditionDate(filters.getDestinationId());
+		} else {
+			remissions = new ArrayList<>();
+		}
 
-	    // Filter by date range only if no other filters have been applied
-	    if (remissions.isEmpty() && filters.getDateInit() != null && filters.getDateEnd() != null) {
-	        remissions = remissionRepository.findByInitAndEndExpeditionDate(filters.getDateInit(), filters.getDateEnd());
-	    }
+		// Filter by date range only if no other filters have been applied
+		if (remissions.isEmpty() && filters.getDateInit() != null && filters.getDateEnd() != null) {
+			remissions = remissionRepository.findByInitAndEndExpeditionDate(filters.getDateInit(),
+					filters.getDateEnd());
+		}
 
-	    remissions.forEach(remission -> {
-	        RemissionVO remissionVO = new RemissionVO();
-	        BeanUtils.copyProperties(remission, remissionVO);
+		remissions.forEach(remission -> {
+			RemissionVO remissionVO = new RemissionVO();
+			BeanUtils.copyProperties(remission, remissionVO);
 
-	        try {
-	            // Streamline the grouping and concatenation of product descriptions
-	            String partidaDescriptions = remission.getProducts().stream()
-	                    .map(Product::getPartidaDescription)
-	                    .distinct()
-	                    .collect(Collectors.joining(", "));
-	            remissionVO.setRemissionPartidas(partidaDescriptions);
-	        } catch (Exception e) {
-	            // Log the exception for better error tracking
-	            e.printStackTrace();
-	        }
+			try {
+				// Streamline the grouping and concatenation of product descriptions
+				String partidaDescriptions = remission.getProducts().stream().map(Product::getPartidaDescription)
+						.distinct().collect(Collectors.joining(", "));
+				remissionVO.setRemissionPartidas(partidaDescriptions);
+			} catch (Exception e) {
+				// Log the exception for better error tracking
+				e.printStackTrace();
+			}
 
-	        remission.getProducts().stream().forEach(
-	        		
-	        		p ->{
-	        			ProductDetailedVO detail = new ProductDetailedVO();
-	        			
-	        			BeanUtils.copyProperties(p,detail);
-	        			BeanUtils.copyProperties(remissionVO,detail);
-	        			productList.add(detail);
-	        			
-	        		}
-	        		);
+			remission.getProducts().stream().forEach(
 
-	    
-	    });
+					p -> {
+						ProductDetailedVO detail = new ProductDetailedVO();
 
-	    return productList;
+						BeanUtils.copyProperties(p, detail);
+						BeanUtils.copyProperties(remissionVO, detail);
+						productList.add(detail);
+
+					});
+
+		});
+
+		return productList;
 	}
-	
-    public RemissionVO checkRemissionDuplication(final String remissionId) {
-        // Ensure the ID is not null or blank before proceeding.
-        if (remissionId == null || remissionId.isBlank()) {
-            throw new IllegalArgumentException("Remission ID cannot be null or blank");
-        }
 
-        RemissionVO remissionVO = new RemissionVO();
-        // Attempt to find the existing Remission.
-        Remission remission = remissionRepository.findById(remissionId)
-                .orElseThrow(() -> new AccessPointException("Origin Remission with given ID does not exist"));
+	public RemissionVO checkRemissionDuplication(final String remissionId) {
+		// Ensure the ID is not null or blank before proceeding.
+		if (remissionId == null || remissionId.isBlank()) {
+			throw new IllegalArgumentException("Remission ID cannot be null or blank");
+		}
 
-        // Check if the status of the Remission is "Activa".
-        if ("Activa".equalsIgnoreCase(remission.getStatus())) {
-            // Update the status and delivered date of the Remission.
-            remission.setStatus("Entregada");
-            remission.setDeliveredDate(new Date());
-            // Save the updated Remission entity to the database.
-            remission = remissionRepository.save(remission);
-            // Copy properties from the Remission entity to the RemissionVO.
-            BeanUtils.copyProperties(remission, remissionVO);
-        } else {
-            throw new AccessPointException("Origin Remission with given ID is not active");
-        }
+		RemissionVO remissionVO = new RemissionVO();
+		// Attempt to find the existing Remission.
+		Remission remission = remissionRepository.findById(remissionId)
+				.orElseThrow(() -> new AccessPointException("Origin Remission with given ID does not exist"));
 
-        return remissionVO;
-    }
-    
+		// Check if the status of the Remission is "Activa".
+		if ("Activa".equalsIgnoreCase(remission.getStatus())) {
+			// Update the status and delivered date of the Remission.
+			remission.setStatus("Entregada");
+			remission.setDeliveredDate(new Date());
+			// Save the updated Remission entity to the database.
+			remission = remissionRepository.save(remission);
+			// Copy properties from the Remission entity to the RemissionVO.
+			BeanUtils.copyProperties(remission, remissionVO);
+		} else {
+			throw new AccessPointException("Origin Remission with given ID is not active");
+		}
+
+		return remissionVO;
+	}
+
 	public RemissionVO remissionCancelation(@Valid RemissionVO remisionRequest) throws ParseException {
 
 		// If the ID is not null or blank, attempt to find the existing Complex.
 		Remission remission = remissionRepository.findById(remisionRequest.getId())
 				.orElseThrow(() -> new AccessPointException("Origin Remission with given ID does not exist"));
 
-		if("Activa".equalsIgnoreCase(remission.getStatus())) {
+		if ("Activa".equalsIgnoreCase(remission.getStatus())) {
 			// Update the existing remision entity with new values.
 			remission.setStatus("Cancelada");
-			
+
 			SimpleDateFormat sdf = new SimpleDateFormat("dd-MM-yyyy HH:mm");
-			
-			if(remisionRequest.getObservations().isEmpty()) {	
-				remission.setObservations(String.format("Remisión cancelada el día %s",sdf.format(new Date())));	
-			}else {
-				remission.setObservations(remisionRequest.getObservations());	
+
+			if (remisionRequest.getObservations().isEmpty()) {
+				remission.setObservations(String.format("Remisión cancelada el día %s", sdf.format(new Date())));
+			} else {
+				remission.setObservations(remisionRequest.getObservations());
 			}
 			// Save the updated Complex entity to the database.
 			remission = remissionRepository.save(remission);
-		}else {
+		} else {
 			throw new AccessPointException("No se puede cancelar la remisión");
 		}
-		
+
 		// Copy properties from the Complex entity back to the ComplexVO.
 		BeanUtils.copyProperties(remission, remisionRequest);
 		return remisionRequest;
 	}
 }
-
